### 2020-07-02 #2426 @whedon
[PRE REVIEW]: OmniCV - A library for Omnidirectional Cameras
**Submitting author:** @kaustubh-sadekar (<a href="http://orcid.org/0000-0001-8895-5319">kaustubh sadekar</a>)
**Repository:** <a href="https://github.com/kaustubh-sadekar/OmniCV-Lib" target ="_blank">https://github.com/kaustubh-sadekar/OmniCV-Lib</a>
**Version:** v1.0
