### 2020-07-03 #2426 @mjsottile
@danielskatz sure.  It might be a couple days to start the process with the holiday weekend.
