### 2020-07-21 #2426 @mjsottile
:wave: @paul92, @innat : Would you be interested in reviewing this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
