### 2020-07-24 #2426 @mjsottile
@kaustubh-sadekar Just an update - I’m having some trouble finding a second reviewer.  This is not unique to your paper - at the moment I’m noticing it’s harder than usual to get responses from folks across reviews, both finding reviewers and getting reviews from people who agreed to review.  I assume it is in part due to the weird times right now and people having generally less capacity than usual.  I’ll keep trying and keep you up to date.
