### 2020-07-27 #2426 @mjsottile
:wave: @ionlights Would you be interested in reviewing this submission for JOSS? Based on your information in the potential-reviewers spreadsheet, this appears to be near your areas of interest.  We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
