+++ paper/paper.bib
@@ -12,12 +12,13 @@ doi="10.1007/978-0-387-31439-6_488",
 url="https://doi.org/10.1007/978-0-387-31439-6_488"
 }
 
-@article{Zhang2016BenefitOL,
-  title={Benefit of large field-of-view cameras for visual odometry},
-  author={Zichao Zhang and Henri Rebecq and Christian Forster and Davide Scaramuzza},
-  journal={2016 IEEE International Conference on Robotics and Automation (ICRA)},
-  year={2016},
-  pages={801-808}
+@inproceedings{Zhang2016BenefitOL,
+author = {Zhang, Zichao and Rebecq, Henri and Forster, Christian and Scaramuzza, Davide},
+year = {2016},
+month = {05},
+pages = {801-808},
+title = {Benefit of large field-of-view cameras for visual odometry},
+doi = {10.1109/ICRA.2016.7487210}
 }
 
 @inproceedings{inproceedings,
@@ -57,7 +58,8 @@ url = {http://papers.nips.cc/paper/6656-learning-spherical-convolution-for-fast-
     year={2018},
     eprint={1811.12633},
     archivePrefix={arXiv},
-    primaryClass={cs.RO}
+    primaryClass={cs.RO},
+    doi = {10.1007/978-3-030-20876-9_3}
 }
 
 
@@ -71,7 +73,8 @@ url = {http://papers.nips.cc/paper/6656-learning-spherical-convolution-for-fast-
   eprinttype = {arXiv},
   eprintclass = {cs.CV},
   note = {{<a href="https://arxiv.org/abs/1807.08957" target="_blank">[arxiv]</a>}},
-  keywords = double-sphere
+  keywords = double-sphere,
+  doi = {10.1109/3DV.2018.00069}
 }
 
 @inproceedings{UCM,
@@ -109,7 +112,8 @@ author={A. {Eichenseer} and A. {Kaup}},
 booktitle={2016 IEEE International Conference on Acoustics, Speech and Signal Processing (ICASSP)},
 title={A data set providing synthetic and real-world fisheye video sequences},
 year={2016},
-volume={},  number={},  pages={1541-1545},}
+volume={},  number={},  pages={1541-1545},
+doi = {10.1109/ICASSP.2016.7471935}}
 
 @misc{py360convert,
 title={py360convert},
