### 2020-08-03 kaustubh-sadekar/OmniCV-Lib#2 @engnadeau
Contribution and authorship
- The [paper](https://github.com/kaustubh-sadekar/OmniCV-Lib/blob/master/paper/paper.md) lists Leena Vachhani and Abhishek Gupta as co-authors. There is no commit history that includes anyone but @kaustubh-sadekar contributing to the code.

How were these authors involved in developing this codebase?

Remember, with JOSS co-authorship is not awarded to supervisors: https://joss.readthedocs.io/en/latest/submitting.html#authorship

cc https://github.com/openjournals/joss-reviews/issues/2533
