+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 ---
 
 # Summary 
-Omnidirectional cameras refer to cameras with large field-of-view (FOV), more than 180°, and ideally 360°. This includes cameras with horizontal 360° field-of-view(FOV) and cameras with FOV spanning a 360° horizontal and more than 90 (ideally 180) vertical field-of-view(FOV). It is beneficial to use such large field-of-view cameras, especially for indoor environments (@Scaramuzza2014). There has been a constant growth in research related to omnidirectional cameras (@Zhang2016BenefitOL). Most of the applications using such cameras either need a mathematical model of the camera or some form of mapping to represent the 360° image information like equirectangular projection or cubemap projection.
+Omnidirectional cameras refer to cameras with large field-of-view (FOV), more than 180°, and ideally 360°. This includes cameras with horizontal 360° field-of-view (FOV) and cameras with FOV spanning a 360° horizontal and more than 90 (ideally 180) vertical field-of-view(FOV). It is beneficial to use such large field-of-view cameras, especially for indoor environments (@Scaramuzza2014). There has been a constant growth in research related to omnidirectional cameras (@Zhang2016BenefitOL). Most of the applications using such cameras either need a mathematical model of the camera or some form of mapping to represent the 360° image information like equirectangular projection or cubemap projection.
 
 OmniCV is a complete library that is written to support and enhance the research related to omnidirectional cameras. It contains, different class methods and APIs to, simulate virtual cameras based on different omnidirectional camera models and also contains methods for interconversion of different representations of a 360° image like equirectangular projection, cubemap projection and perspective projection. The library also contains some essential GUI based software written using the package APIs that enable the user to determine various parameters of the camera used to capture any given image. All the functionalities are written in C++ as well as python. The operations are optimized to ensure significant use in real-time applications. Moreover, ROS nodes for all the functionalities are also provided for easy integration with any existing ROS project. 
 
