### 2020-08-03 kaustubh-sadekar/OmniCV-Lib#5 @engnadeau
Syntax
- bad comma in https://github.com/kaustubh-sadekar/OmniCV-Lib/blob/master/paper/paper.md : ` to, simulate virtual `
- other minor syntax issues

cc https://github.com/openjournals/joss-reviews/issues/2533
