### 2020-08-04 kaustubh-sadekar/OmniCV-Lib#2 @kaustubh-sadekar
There were several active non-code contributions from both the CO-authors towards shaping the OmniCV library. Their active direction in terms of designing the architecture of the library, deciding various functionalities of OmniCV that would be useful for robotics research based on useful reference, deciding the architecture of ROS nodes, developing the design and flow of the entire OmniCV documentation as well as the OmniCV paper. Hence active contributions from both the CO-authors have a significant role in transforming OmniCV from a small piece of unstructured code to a complete library with several useful features.
