### 2020-08-05 kaustubh-sadekar/OmniCV-Lib#2 @engnadeau
@mjsottile , i'll leave it to you as the editor to determine the suitability of co-authorship with respect to https://joss.readthedocs.io/en/latest/submitting.html#authorship
