### 2020-08-05 kaustubh-sadekar/OmniCV-Lib#2 @mjsottile
Based on what @kaustubh-sadekar stated about their role, their contributions sound like they fit the requirements stated in the JOSS documents - active direction and non-code contributions are acceptable for inclusion as co-authors.  There is no requirement that all authors appear in the git commit history.  Their co-authorship role looks fine to me.
