### 2020-08-13 2b84cd0
Update paper.md
