### 2020-08-13 kaustubh-sadekar/OmniCV-Lib#5 @kaustubh-sadekar
Thank you very much for the suggestions. I have updated paper.md file.
