### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#8 @engnadeau
PyPI README is empty
Please update your packaging settings to ensure that [PyPI has a README](https://pypi.org/project/omnicv/)

![image](https://user-images.githubusercontent.com/6395915/90992170-eab9ba80-e57b-11ea-8903-ed0898848ccd.png)

cc https://github.com/openjournals/joss-reviews/issues/2533
