### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#9 @engnadeau
Python requirements and installation instructions
- The python [requirements](https://github.com/kaustubh-sadekar/OmniCV-Lib/raw/master/requirements.txt) are currently fixed (i.e., `==`)
- Packages typically specify minimum requirements, else a `pip install` will have conflicts with the globally installed packages
- Moreover, the installation instructions should also include notes about installing this package into a virtualenv (e.g., using `pipenv` or `poetry`). This is the ideal method for scientific research, as it freezes the requirements and isolates the environment

cc https://github.com/openjournals/joss-reviews/issues/2533
