+++ setup.py
@@ -1,3 +1,8 @@
 import setuptools
 
-setuptools.setup(name = 'omnicv', packages = ['omnicv'], version = '1.0', author = 'Kaustubh Sadekar, Leena Vachhani, Abhishek Gupta')
+setuptools.setup(
+    name="omnicv",
+    packages=["omnicv"],
+    version="1.0",
+    author="Kaustubh Sadekar, Leena Vachhani, Abhishek Gupta",
+)
