+++ applications/set_fisheye_params_gui.py
@@ -7,8 +7,10 @@ from omnicv import fisheyeImgConv
 
 # cap = cv2.VideoCapture("vid1.mp4")
 
+
 def nothing(x):
     pass
+
 
 WINDOW_NAME = "image"
 Image_Path = sys.argv[1]
@@ -16,22 +18,22 @@ RADIUS = None
 
 # Example of using the converter class
 frame = cv2.imread(Image_Path)
-cv2.namedWindow(WINDOW_NAME,cv2.WINDOW_NORMAL)
-cv2.resizeWindow(WINDOW_NAME,800,800)
-N = 2*max(frame.shape)
-cv2.createTrackbar('radius',WINDOW_NAME,0,N,nothing)
-cv2.createTrackbar('Cx',WINDOW_NAME,N//2,N,nothing)
-cv2.createTrackbar('Cy',WINDOW_NAME,N//2,N,nothing)
+cv2.namedWindow(WINDOW_NAME, cv2.WINDOW_NORMAL)
+cv2.resizeWindow(WINDOW_NAME, 800, 800)
+N = 2 * max(frame.shape)
+cv2.createTrackbar("radius", WINDOW_NAME, 0, N, nothing)
+cv2.createTrackbar("Cx", WINDOW_NAME, N // 2, N, nothing)
+cv2.createTrackbar("Cy", WINDOW_NAME, N // 2, N, nothing)
 
 while True:
     if True:
         frame = cv2.imread(Image_Path)
-        radius = cv2.getTrackbarPos('radius',WINDOW_NAME)
-        Cx = cv2.getTrackbarPos('Cx',WINDOW_NAME)
-        Cy = cv2.getTrackbarPos('Cy',WINDOW_NAME)
-        frame = cv2.circle(frame,(Cx,Cy),radius,(0,200,0),2)
-        cv2.imshow(WINDOW_NAME,frame)
-        if cv2.waitKey(1)&0xFF == 27:
+        radius = cv2.getTrackbarPos("radius", WINDOW_NAME)
+        Cx = cv2.getTrackbarPos("Cx", WINDOW_NAME)
+        Cy = cv2.getTrackbarPos("Cy", WINDOW_NAME)
+        frame = cv2.circle(frame, (Cx, Cy), radius, (0, 200, 0), 2)
+        cv2.imshow(WINDOW_NAME, frame)
+        if cv2.waitKey(1) & 0xFF == 27:
             RADIUS = radius
             cv2.destroyAllWindows()
             break
@@ -39,43 +41,77 @@ while True:
 
 WINDOW_NAME = "set aperture"
 
-cv2.namedWindow(WINDOW_NAME,cv2.WINDOW_NORMAL)
-cv2.resizeWindow(WINDOW_NAME,600,1200)
-cv2.createTrackbar('aperture',WINDOW_NAME,0,1000,nothing)
-cv2.createTrackbar('del Cx',WINDOW_NAME,500,1000,nothing)
-cv2.createTrackbar('del Cy',WINDOW_NAME,500,1000,nothing)
+cv2.namedWindow(WINDOW_NAME, cv2.WINDOW_NORMAL)
+cv2.resizeWindow(WINDOW_NAME, 600, 1200)
+cv2.createTrackbar("aperture", WINDOW_NAME, 0, 1000, nothing)
+cv2.createTrackbar("del Cx", WINDOW_NAME, 500, 1000, nothing)
+cv2.createTrackbar("del Cy", WINDOW_NAME, 500, 1000, nothing)
 
 # Example of using the converter class
 frame = cv2.imread(Image_Path)
[…]

[…]
