### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#12 @engnadeau
Automated CI
There is currently no automated testing and deployment of the codebase. For this to be considered, a CI/CD setup (e.g., GitHub Actions) should be created that tests PRs, creates releases, and deploys to PyPI

cc https://github.com/openjournals/joss-reviews/issues/2533
