### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#14 @engnadeau
Simplify entrypoint
When getting started with the repo, the two primary entrypoints are "building the code" and "running examples".

While a skilled developer will have no problem building with cmake, running `chmod +x tests/test.sh`, and understanding what is needed to run the examples, these are an extra impedance for someone new to the repo.

I'd recommend using a `Makefile` as the entrypoint for building, running tests, etc. The READMEs will need to be updated as well.

cc https://github.com/openjournals/joss-reviews/issues/2533
