### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#15 @engnadeau
Shebang vs instructions
The README instructions call for `bash test.sh` but the shebang in `test.sh` calls for `sh`. These are not the same. Moreover, calling `bash` on a file isn't standard practice. Rather the permissions should be updated so that we call the file like this: `./test.sh` (see #7 )

cc https://github.com/openjournals/joss-reviews/issues/2533
