### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#15 @engnadeau
To be honest, `test.sh` should be fully replaced by a `Makefile` and the tests should be called using `make test` and `make test-gui` commands. See #14
