### 2020-08-24 kaustubh-sadekar/OmniCV-Lib#16 @engnadeau
Testing doesn't use testing frameworks
All the tests, whether C++ or Python, aren't using modern, tested, and robust testing frameworks. For Python, I personally recommend `pytest`. For C++, I personally recommend `Catch2`.

This should also be integrated with #12

cc https://github.com/openjournals/joss-reviews/issues/2533
