### 2020-08-26 kaustubh-sadekar/OmniCV-Lib#16 @kaustubh-sadekar
Thank you so much for all the suggestions and guidance. It's my first JOSS paper and it's a great learning experience for me. I will definitely improve the test process keeping all the suggestions in mind as early as possible.
