### 2020-08-29 71f3d14
Updated test files
Test files now support automated testing. For python, pytest is used and for C++, catch2 has been used.
