+++ omnicv/omnicv.py
@@ -61,6 +61,7 @@ class fisheyeImgConv:
         y = self.radius*np.cos((i*1.0/self.Hd - 0.5)*np.pi)*np.sin((j*1.0/self.Hd - 0.5)*np.pi)
         z = self.radius * np.sin((i*1.0/self.Hd-0.5)*np.pi)
 
+        # indx = np.logical_not(np.arccos(y/np.sqrt(x**2+y**2+z**2))/np.pi*180 > self.aperture/2)
         r = 2*np.arctan2(np.sqrt(x**2+z**2), y)/np.pi*180/self.aperture*self.radius
         theta = np.arctan2(z,x)
 
@@ -147,6 +148,8 @@ class fisheyeImgConv:
         self.map_x = eqrec_x
         self.map_y = eqrec_y
 
+        # print(self.map_x.shape)
+        # print(self.map_y.shape)
         
         dstFrame = cv2.remap(srcFrame,self.map_x,self.map_y,interpolation=cv2.INTER_LINEAR, borderMode=cv2.BORDER_CONSTANT)
 
@@ -170,6 +173,8 @@ class fisheyeImgConv:
 
         	srcFrame = np.hstack((pZ,cv2.flip(pX,1),cv2.flip(nZ,1),nX,cv2.flip(pY,0),nY))
 
+        # cv2.imshow("src frame",srcFrame)
+        # cv2.waitKey(0)
 
         inShape = srcFrame.shape[:2]
         self.Hd = outShape[0]
@@ -197,7 +202,11 @@ class fisheyeImgConv:
         idx = h // 2 - np.round(np.arctan(np.cos(idx)) * h / np.pi).astype(int)
         for i, j in enumerate(idx):
             mask[:j, i] = 1
+        # cv2.imshow("mask first",mask.astype(np.uint8)*90)
+        # cv2.waitKey(0)
         mask = np.roll(mask,w//8,1)
+        # cv2.imshow("mask first",mask.astype(np.uint8)*90)
+        # cv2.waitKey(0)
         mask = np.concatenate([mask] * 4, 1)
 
         tp[mask] = 4
@@ -205,6 +214,8 @@ class fisheyeImgConv:
 
         tp = tp.astype(np.int32)
 
+        # cv2.imshow("mask",tp.astype(np.uint8)*60)
+        # cv2.waitKey(0)
 
         coor_x = np.zeros((h, w))
         coor_y = np.zeros((h, w))
@@ -253,6 +264,8 @@ class fisheyeImgConv:
 
         equ_h, equ_w = img.shape[:2]
 
+        # equ_cx = (equ_w - 1) / 2.0
+        # equ_cy = (equ_h - 1) / 2.0
         equ_cx = (equ_w) / 2.0
         equ_cy = (equ_h) / 2.0
 
@@ -263,10 +276,12 @@ class fisheyeImgConv:
         c_y = (Hd) / 2.0
 
         wangle = (180 - wFOV) / 2.0
+        # w_len = 2 * 1 * np.sin(np.radians(wFOV / 2.0))/np.cos(np.radians(wFOV/2.0))
         w_len = 2 * np.tan(np.radians(wFOV / 2.0))
         w_interval = w_len / (Wd)
 
         hangle = (180 - hFOV) / 2.0
+        # h_len = 2 * 1 * np.sin(np.radians(hFOV / 2.0))/np.cos(np.radians(hFOV/2.0))
         h_len = 2 *np.tan(np.radians(hFOV / 2.0))
         h_interval = h_len / (Hd)
 
@@ -310,9 +325,12 @@ class fisheyeImgConv:
[…]
