+++ omnicv/imgMap.cpp
@@ -1,0 +1,957 @@
+#include"utils.hpp"
+#include<fstream>
+#include<iostream>
+#include<string>
+#include <complex> 
+#define CV_INTER_CUBIC 2
+
+fisheyeImgConv::fisheyeImgConv(std::string paramFilePath)
+: filePath {paramFilePath} {}
+
+void fisheyeImgConv::meshgrid(const cv::Range &xgv, const cv::Range &ygv,cv::Mat &X, cv::Mat &Y)
+{
+  std::vector<int> t_x, t_y;
+  for (int i = xgv.start; i <= xgv.end; i++) t_x.push_back(i);
+  for (int i = ygv.start; i <= ygv.end; i++) t_y.push_back(i);
+
+  cv::repeat(cv::Mat(t_x).reshape(1,1), cv::Mat(t_y).total(), 1, X);
+  cv::repeat(cv::Mat(t_y).reshape(1,1).t(), 1, cv::Mat(t_x).total(), Y);
+}
+
+cv::Mat fisheyeImgConv::RMat(double alpha, double beta, double gamma)
+{	
+	cv::Mat R;
+
+	cv::Mat Rx = (cv::Mat_<double>(3,3) << 1,0,0,0,std::cos(alpha),-1*std::sin(alpha),0,std::sin(alpha),std::cos(alpha));
+	cv::Mat Ry = (cv::Mat_<double>(3,3) << std::cos(beta),0,1*std::sin(beta),0,1,0,-1*std::sin(beta),0,std::cos(beta));
+	cv::Mat Rz = (cv::Mat_<double>(3,3) << std::cos(gamma),-1*std::sin(gamma),0,std::sin(gamma),std::cos(gamma),0,0,0,1);
+	R = Rz*Ry*Rx;
+
+	return R;
+}
+void fisheyeImgConv::fisheye2equirect(const cv::Mat &srcFrame, cv::Mat &dstFrame, cv::Size outShape, int aperture, int delx, int dely, int radius, bool edit_mode)
+{
+	this->Hs = srcFrame.rows;
+	this->Ws = srcFrame.cols;
+	this->Hd = outShape.height;
+	this->Wd = outShape.width;
+
+	this->map_x = cv::Mat::zeros(this->Hd,this->Wd,CV_32FC1);
+	this->map_y = cv::Mat::zeros(this->Hd,this->Wd,CV_32FC1);
+
+	this->Cx = int(this->Ws/2 - delx);
+	this->Cy = int(this->Hs/2 - dely);
+
+
+	if (!(edit_mode))
+	{
+		std::fstream f;
+		std::string val;
+		int temp;
+
+		f.open(this->filePath,ios::in);
+		if(!f)
+			std::cout << "File path wrong! ";
+		else
+		{
+			std::getline(f, val);
+			this->radius = stoi(val);
+			std::getline(f, val);
+			this->aperture = stoi(val);
+			std::getline(f, val);
+			delx = stoi(val);
+			std::getline(f, val);
+			dely = stoi(val);
+		}
+	}
+
+	float x,y,z,r,theta;
+
+	for(size_t i{0}; i<this->Hd; i++)
+	{
+		float* dataX = this->map_x.ptr<float>(i);
+		float* datay = this->map_y.ptr<float>(i);
+
+		for(size_t j{0}; j<this->Wd; j++)
+		{
+			x = this->radius*std::cos((i*1.0/this->Hd-0.5)*CV_PI)*std::cos((j*1.0/this->Hd-0.5)*CV_PI);
+			y = this->radius*std::cos((i*1.0/this->Hd-0.5)*CV_PI)*std::sin((j*1.0/this->Hd-0.5)*CV_PI);
[…]
