### 2020-08-29 fa5a2f3
Add Makefile for easy setup
Makefile is written to ensure easy building and testing
