### 2020-08-30 #2533 @mjsottile
:wave: @ionlights - you had mentioned that you were busy until mid-August when we opened the review, so I assume you've been busy with your pre-existing commitment.  Does it look like you'll have capacity to handle the review in the not-too-distant future?  No pressure to rush - I'm just checking to make sure you're still interested and available.
