### 2020-09-13 kaustubh-sadekar/OmniCV-Lib#8 @kaustubh-sadekar
@nnadeau I have added README to the PyPi. Here is the link for your kind reference - https://pypi.org/project/omnicv/
