### 2020-09-13 kaustubh-sadekar/OmniCV-Lib#16 @kaustubh-sadekar
I have updated the test files based on your suggestions.
