### 2020-09-13 kaustubh-sadekar/OmniCV-Lib#15 @kaustubh-sadekar
@nnadeau based on your valuable suggestions I have created a single Makefile and have provided several options to the user to build and run the tests. All the options have been explained in the installation instructions here : https://pypi.org/project/omnicv/
