### 2020-09-13 kaustubh-sadekar/OmniCV-Lib#12 @kaustubh-sadekar
I have completed writing tests using standard testing frameworks based on your suggestions. I should be able to complete the CI/ CD setup within this week.
