+++ omnicv/Makefile
@@ -31,6 +31,12 @@ test-py:
 	@echo "=============================================== \n"
 	cd ../tests/ && ${PYTHON} -W ignore -m pytest --ignore=./test_visual.py
 
+test-py-gui:
+	@echo "=============================================== \n"
+	@echo "Running visual tests for Python extension \n"
+	@echo "=============================================== \n"
+	cd ../tests/ && ${PYTHON} -W ignore -m pytest -s --ignore=./test_.py && ./test
+	
 build-cpp:
 	@echo "=============================================== \n"
 	@echo "Building omnicv c++ files"
@@ -44,6 +50,12 @@ test-cpp:
 	@echo "=============================================== \n"
 	cd ../tests/ && ./test
 
+test-cpp-gui:
+	@echo "=============================================== \n"
+	@echo "Running visual tests for C++ extension \n"
+	@echo "=============================================== \n"
+	cd ../tests/ && ./test_visual
+
 build-all:
 	$(MAKE) build-py
 	$(MAKE) build-cpp
@@ -52,22 +64,6 @@ test-all:
 	$(MAKE) test-py
 	$(MAKE) test-cpp
 
-build:	
-	@echo "=============================================== \n"
-	@echo "Installing omnicv python library from source"
-	@echo "=============================================== \n"
-	cd ../ && sudo ${PYTHON} setup.py install
-	@echo "=============================================== \n"
-	@echo "Building omnicv c++ files"
-	@echo "=============================================== \n"
-	cd ../build/ && cmake -Wno-dev CMakeLists.txt && $(MAKE)
-	@echo "=============================================== \n"
-	@echo "Running tests \n"
-	@echo "=============================================== \n"
-	$(MAKE) test
-
-test:
-	cd ../tests/ && ${PYTHON} -W ignore -m pytest --ignore=./test_visual.py && ./test
-
-test-gui:
-	cd ../tests/ && ${PYTHON} -W ignore -m pytest -s --ignore=./test_.py && ./test
+test-all-gui:
+	$(MAKE) test-py-gui
+	$(MAKE) test-cpp-gui
