### 2020-09-26 kaustubh-sadekar/OmniCV-Lib#16 @kaustubh-sadekar
I have further modified the testing files and used `pytest` and `Catch2`. Please let me know if it is as per requirements and if we could close the issue.

cc openjournals/joss-reviews#2533
