### 2020-09-26 kaustubh-sadekar/OmniCV-Lib#14 @kaustubh-sadekar
@nnadeau I would like to kindly convey to you that the issue has been resolved by creating Makefiles and also updating the documentation accordingly. Kindly provide your further reviews so we could decide if the thread could be closed.
