+++ requirements.txt
@@ -3,7 +3,6 @@ importlib-metadata>=2.0.0
 iniconfig>=1.0.1
 more-itertools>=8.5.0
 numpy>=1.18.5
-omnicv>=1.1.0
 opencv-python>=4.4.0.42
 packaging>=20.4
 pathlib2>=2.3.5
