+++ omnicv/__init__.py
@@ -1,1 +1,2 @@
+# pyflakes.ignore
 from omnicv.omnicv import *
