+++ omnicv/omnicv.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env/python
 import cv2
 import numpy as np
-import math
-import time
 
 
-def rmat(alpha, beta, gamma):
+def rmat(alpha,
+         beta,
+         gamma):
+
     rx = np.array(
         [
             [1, 0, 0],
@@ -32,7 +33,9 @@ def rmat(alpha, beta, gamma):
 
 
 class fisheyeImgConv:
-    def __init__(self, param_file_path=None):
+    def __init__(self,
+                 param_file_path=None
+                 ):
 
         self.Hd = None
         self.Wd = None
@@ -41,9 +44,15 @@ class fisheyeImgConv:
         self.singleLens = False
         self.filePath = param_file_path
 
-    def fisheye2equirect(
-        self, srcFrame, outShape, aperture=0, delx=0, dely=0, radius=0, edit_mode=False
-    ):
+    def fisheye2equirect(self,
+                         srcFrame,
+                         outShape,
+                         aperture=0,
+                         delx=0,
+                         dely=0,
+                         radius=0,
+                         edit_mode=False):
+
         inShape = srcFrame.shape[:2]
         self.Hs = inShape[0]
         self.Ws = inShape[1]
@@ -77,7 +86,8 @@ class fisheyeImgConv:
             dely = int(f.readline())
             f.close
 
-        i, j = np.meshgrid(np.arange(0, int(self.Hd)), np.arange(0, int(self.Wd)))
+        i, j = np.meshgrid(np.arange(0, int(self.Hd)),
+                           np.arange(0, int(self.Wd)))
         xyz = np.zeros((self.Hd, self.Wd, 3))
         x, y, z = np.split(xyz, 3, axis=-1)
 
@@ -93,7 +103,6 @@ class fisheyeImgConv:
         )
         z = self.radius * np.sin((i * 1.0 / self.Hd - 0.5) * np.pi)
 
-        # indx = np.logical_not(np.arccos(y/np.sqrt(x**2+y**2+z**2))/np.pi*180 > self.aperture/2)
         r = (
             2
             * np.arctan2(np.sqrt(x ** 2 + z ** 2), y)
@@ -104,8 +113,10 @@ class fisheyeImgConv:
         )
         theta = np.arctan2(z, x)
 
-        self.map_x = np.multiply(r, np.cos(theta)).T.astype(np.float32) + self.Cx
-        self.map_y = np.multiply(r, np.sin(theta)).T.astype(np.float32) + self.Cy
+        self.map_x = np.multiply(r, np.cos(theta)).T.astype(np.float32)
+        + self.Cx
+        self.map_y = np.multiply(r, np.sin(theta)).T.astype(np.float32)
+        + self.Cy
 
         return cv2.remap(
             srcFrame,
@@ -115,7 +126,12 @@ class fisheyeImgConv:
             borderMode=cv2.BORDER_CONSTANT,
         )
 
-    def equirect2cubemap(self, srcFrame, side=256, modif=False, dice=False):
+    def equirect2cubemap(self,
+                         srcFrame,
+                         side=256,
+                         modif=False,
+                         dice=False
+                         ):
 
         self.dice = dice
         self.side = side
@@ -129,115 +145,117 @@ class fisheyeImgConv:
             -1,
         )
 
-        # Creating a matrix that contains x,y,z values of all 6 faces of a unit cube
[…]
