+++ omnicv/requirements.txt
@@ -1,0 +1,16 @@
+attrs==20.2.0
+importlib-metadata==2.0.0
+iniconfig==1.0.1
+more-itertools==8.5.0
+numpy==1.18.5
+omnicv==1.1.0
+opencv-python==4.4.0.42
+packaging==20.4
+pathlib2==2.3.5
+pluggy==0.13.1
+py==1.9.0
+pyparsing==2.4.7
+pytest==6.0.2
+six==1.15.0
+toml==0.10.1
+zipp==1.2.0
