+++ Examples/fisheye2equirect.py
@@ -1,9 +1,6 @@
 #!/usr/bin/env/python
 import cv2
-import numpy as np
-import math
 import sys
-import time
 from omnicv import fisheyeImgConv
 
 Img_path = sys.argv[1]
@@ -19,7 +16,7 @@ cv2.destroyAllWindows()
 outShape = [200, 400]
 inShape = frame.shape[:2]
 
-########################### Uncomment the block to run a perticular experiment  ##########################
+# Uncomment the block to run a perticular experiment
 
 """
 # In case of fisheye lens placed vertically
@@ -28,9 +25,13 @@ mapper = fisheyeImgConv(param_file_path)
 frame2 = mapper.fisheye2equirect(frame,outShape)
 conv_time = time.time() - start
 start = time.time()
-# Use the below line if there are multiple images and the mapping is not changing in case of a video
+
+# Use the below line if there are multiple images
+# and the mapping is not changing in case of a video
+
 for i in range(10):
-	frame2 = mapper.applyMap(0,frame)
+    frame2 = mapper.applyMap(0,frame)
+
 remap_time = (time.time() - start)/10.0
 total_time = conv_time + remap_time
 print("time consumed :",total_time)
@@ -43,7 +44,7 @@ cv2.waitKey(0)
 
 # """
 # In case of fisheye lens placed horizontally
-# NOTE : This conversion is really helpful when we look at the perspective view.
+# Note: This conversion is really helpful when we look at the perspective view.
 mapper = fisheyeImgConv(param_file_path)
 frame2 = mapper.fisheye2equirect(frame, outShape)
 frame2 = mapper.equirect2cubemap(frame2, modif=True)
