+++ tests/srcpath.py
@@ -1,0 +1,6 @@
+import os
+import sys
+
+modpath = os.getcwd()+"/omnicv/"
+sys.path.insert(0, modpath)
+print(modpath)
