+++ tests/test_.py
@@ -1,5 +1,4 @@
-#!/usr/bin/env/python
-
+import srcpath # noqa
 import cv2
 import numpy as np
 import math
