+++ tests/test.sh
@@ -1,12 +1,0 @@
-#!/bin/sh
-
-if [ $1 == "1" ]
-then
-	python3 -W ignore tests.py
-	./test
-	python3 -W ignore tests_visual.py
-	./test_visual
-else
-	python3 -W ignore tests.py
-	./test
-fi
