+++ tests/srcpath.py
@@ -1,6 +1,5 @@
 import os
 import sys
 
-modpath = os.getcwd()+"/omnicv/"
+modpath = "../omnicv/"
 sys.path.insert(0, modpath)
-print(modpath)
