+++ README.md
@@ -2,6 +2,12 @@
   <img width="250" src="logo.png">
 </p>
 
+
+<p align="center">
+  <a href="https://github.com/kaustubh-sadekar/OmniCV-Lib/actions?query=workflow%3Apython-tests"><img alt="python-tests" src="https://github.com/kaustubh-sadekar/OmniCV-Lib/workflows/python-tests/badge.svg"></a>
+  <a href="https://badge.fury.io/py/omnicv"><img src="https://badge.fury.io/py/omnicv.svg" alt="PyPI version" height="18"></a>
+</p>
+
 A **computer vision library for omnidirectional(360&deg;) cameras**. This package is divided into two parts:
 * **Basic functions** for inter-conversion of different types of mappings associated with omni directional cameras, virtual reality and 360&deg; videos, like cubemap, spherical projections, perspective projection and equirectangular projection.
 * **Software applications** like **360&deg; video viewer**, **fisheye image generator** with variable intrinsic properties, GUI to determine **fisheye camera paraeters**.
