+++ .github/workflows/testsCpp.yml
@@ -11,8 +11,4 @@ jobs:
     - name: Install OpenCV
       uses: ./ # Uses an action in the root directory
       with:
-        opencv-version: '4.0.0'
-    - name: build omnicv
-      run: |
-       echo pwd
-      
+        opencv-version: '4.0.0'  
