+++ .github/workflows/testsCpp.yml
@@ -15,7 +15,7 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd ~ 
+       cd
        pwd
        cd omnicv
        make build-cpp
