+++ .github/workflows/testsCpp.yml
@@ -15,7 +15,7 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd
+       cd ../../
        pwd
+       ls
        cd omnicv
-       make build-cpp
