+++ .github/workflows/testsCpp.yml
@@ -15,5 +15,5 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd ../../
-       ls OmniCV-Lib/
+       cd ../../OmniCV-Lib/
+       ls
