+++ .github/workflows/testsCpp.yml
@@ -15,6 +15,6 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd ~
+       cd 
        cd ../../OmniCV-Lib/
        ls
