+++ .github/workflows/testsCpp.yml
@@ -15,6 +15,5 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd 
        cd ../../OmniCV-Lib/
        ls
