+++ .github/workflows/testsCpp.yml
@@ -15,8 +15,5 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       pwd
-       cd ~
-       pwd
-       cd ../../OmniCV-Lib/
+       cd /home/runner/work/OmniCV-Lib/OmniCV-Lib
        ls
