+++ .github/workflows/testsCpp.yml
@@ -15,5 +15,5 @@ jobs:
     - uses: actions/checkout@v2
     - name: Install dependencies
       run: |
-       cd /home/runner/work/OmniCV-Lib/OmniCV-Lib
+       cd /home/runner/work/OmniCV-Lib/OmniCV-Lib/omnicv
        ls
