### 2020-09-30 kaustubh-sadekar/OmniCV-Lib#12 @kaustubh-sadekar
The repository now supports automated testing and deployment. I have created all the CI/CD workflows as per your suggestion. Kindly let me know if further modifications are required or it is fine to close the issue tracker.
