+++ .github/workflows/testsCpp.yml
@@ -1,4 +1,4 @@
-name: cpp-build
+name: cpp-build-test
 
 on:
   push:
