### 2020-10-02 52a4b95
Delete test_visual.cpp
