+++ tests/test_visual.cpp
@@ -1,163 +1,0 @@
-#include<iostream>
-#include<opencv2/opencv.hpp>
-#include"../omnicv/utils.hpp"
-#include <opencv2/core/core.hpp>
-
-int main()
-{
-	fisheyeImgConv mapper1;
-
-	std::cout << "\n\n########################################################\n";
-	std::cout << "Running visual tests for C++ library ...." << std::endl;
-
-	cv::namedWindow("output image",cv::WINDOW_NORMAL);
-	cv::resizeWindow("output image",400,400);
-	cv::namedWindow("reference image",cv::WINDOW_NORMAL);
-	cv::resizeWindow("reference image",400,400);
-
-	std::string img_path = "./input/equirect.jpg";
-
-	cv::Mat equiRect,fisheye;
-
-	equiRect = cv::imread(img_path);
-	assert(equiRect.cols != 0 && "Wrong path... Run the tests.py code after changing the directory inside test folder");
-
-	std::cout << "[1] Testing equirect2Fisheye_UCM method ..."<<std::endl;
-	mapper1.equirect2Fisheye_UCM(equiRect,fisheye, cv::Size (250,250),50,0.5);
-	img_path = "./outputs/UCM_out.jpg";
-	cv::Mat refImg = cv::imread(img_path);
-	cv::imshow("output image",fisheye);
-	cv::imshow("reference image",refImg);
-	cv::waitKey(0);
-	assert (refImg.cols !=0 && "Output reference image not present");
-	assert ((fisheye.cols==250) && (fisheye.rows==250) &&"Output dimensions for generated fisheye image do not match !");
-
-	std::cout << "[2] Testing equirect2Fisheye_EUCM method ..." << std::endl;
-	mapper1.equirect2Fisheye_EUCM(equiRect,fisheye, cv::Size (250,250),100,0.4,2);
-	img_path = "./outputs/EUCM_out.jpg";
-	refImg = cv::imread(img_path);
-	cv::imshow("output image",fisheye);
-	cv::imshow("reference image",refImg);
-	cv::waitKey(0);
-	assert (refImg.cols !=0 && "Output reference image not present");
-	assert ((fisheye.cols==250) && (fisheye.rows==250) &&"Output dimensions for generated fisheye image do not match !");
-
-	std::cout << "[3] Testing equirect2Fisheye_FOV method ..." << std::endl;
-	mapper1.equirect2Fisheye_FOV(equiRect,fisheye, cv::Size (250,250),40,0.5);
-	img_path = "./outputs/FOV_out.jpg";
-	refImg = cv::imread(img_path);
-	cv::imshow("output image",fisheye);
-	cv::imshow("reference image",refImg);
-	cv::waitKey(0);
-	assert (refImg.cols !=0 && "Output reference image not present");
-	assert ((fisheye.cols==250) && (fisheye.rows==250) &&"Output dimensions for generated fisheye image do not match !");
-
-	std::cout << "[4] Testing equirect2Fisheye_DS method ..." << std::endl;
-	mapper1.equirect2Fisheye_DS(equiRect,fisheye, cv::Size (250,250),90,0.4,0.8);
-	img_path = "./outputs/DS_out.jpg";
-	refImg = cv::imread(img_path);
-	cv::imshow("output image",fisheye);
-	cv::imshow("reference image",refImg);
-	cv::waitKey(0);
-	assert (refImg.cols !=0 && "Output reference image not present");
-	assert ((fisheye.cols==250) && (fisheye.rows==250) &&"Output dimensions for generated fisheye image do not match !");
-
-	std::string params_file_path = "../fisheyeParams.txt";
-	
-	mapper1.filePath = "./input/fisheyeParams.txt";
-
-	img_path = "./input/fisheye.jpg";
-	fisheye = cv::imread(img_path);
[…]
