### 2020-10-02 a5418b5
Delete test_visual.py
