+++ tests/test_visual.py
@@ -1,178 +1,0 @@
-#!/usr/bin/env/python
-
-import cv2
-import numpy as np
-import math
-import time
-import sys
-import os
-from omnicv import fisheyeImgConv
-
-print("\n\n########################################################\n")
-print("Running visual tests for python library ....")
-print("\n\n")
-input("Visual tests need user input for verification ...")
-
-cv2.namedWindow("output image and reference image",0)
-cv2.resizeWindow("output image and reference image",800,400)
-
-img_path = "./input/equirect.jpg"
-def test_inputPath1():
-    assert os.path.isfile(img_path),"Wrong path... Run the tests.py code after changing the directory inside test folder"
-
-param_file_path = "./input/fisheyeParams.txt"
-def test_parameterFilePath():
-    assert os.path.isfile(param_file_path),"fisheyeParams.txt file not present"
-
-mapper = fisheyeImgConv(param_file_path)
-equiRect = cv2.imread(img_path)
-#inShape = equiRect.shape[:2]
-
-#print("[1] Testing equirect2Fisheye_UCM method ...")
-def test_equirect2Fisheye_UCM():
-    fisheye = mapper.equirect2Fisheye_UCM(equiRect,outShape=[250,250],xi=0.5)
-    img_path = "./outputs/UCM_out.jpg"
-    cv2.imshow("output image and reference image",np.hstack((fisheye,cv2.imread(img_path))));
-    cv2.waitKey(0)
-    ok = input("enter y if output and reference image match")
-    assert ok.lower() == 'y', "Visual test failure reported by user for equirect2Fisheye_UCM method"
-    assert os.path.isfile(img_path),"Output reference image not present"
-    assert fisheye.shape[:2]==(250,250),"Output dimensions for generated fisheye image do not match !"
-"""
-print("[2] Testing equirect2Fisheye_EUCM method ...")
-fisheye = mapper.equirect2Fisheye_EUCM(equiRect,outShape=[250,250],f=100,a_=0.4,b_=2,angles=[0,0,0])
-img_path = "./outputs/EUCM_out.jpg"
-cv2.imshow("output image and reference image",np.hstack((fisheye,cv2.imread(img_path))));
-cv2.waitKey(0)
-assert os.path.isfile(img_path),"Output reference image not present"
-assert fisheye.shape[:2]==(250,250),"Output dimensions for generated fisheye image do not match !"
-
-print("[3] Testing equirect2Fisheye_FOV method ...")
-fisheye = mapper.equirect2Fisheye_FOV(equiRect,outShape=[250,250],f=40,w_=0.5,angles=[0,0,0])
-img_path = "./outputs/FOV_out.jpg"
-cv2.imshow("output image and reference image",np.hstack((fisheye,cv2.imread(img_path))));
-cv2.waitKey(0)
-assert os.path.isfile(img_path),"Output reference image not present"
-assert fisheye.shape[:2]==(250,250),"Output dimensions for generated fisheye image do not match !"
-
-print("[4] Testing equirect2Fisheye_DS method ...")
-fisheye = mapper.equirect2Fisheye_DS(equiRect,outShape=[250,250],f=90,a_=0.4,xi_=0.8,angles=[0,0,0])
-img_path = "./outputs/DS_out.jpg"
-cv2.imshow("output image and reference image",np.hstack((fisheye,cv2.imread(img_path))));
-cv2.waitKey(0)
-assert os.path.isfile(img_path),"Output reference image not present"
-assert fisheye.shape[:2]==(250,250),"Output dimensions for generated fisheye image do not match !"
-
-img_path = "./input/fisheye.jpg"
[…]
