### 2020-10-02 a1c0803
Delete tests.py
