+++ tests/tests.py
@@ -1,191 +1,0 @@
-#!/usr/bin/env/python
-
-import cv2
-import numpy as np
-import math
-import time
-import sys
-import os
-from omnicv import fisheyeImgConv
-
-mapper = fisheyeImgConv()
-
-print("\n\n########################################################\n")
-print("Running tests for python library ....")
-
-
-img_path = "./input/equirect.jpg"
-assert os.path.isfile(
-    img_path
-), "Wrong path... Run the tests.py code after changing the directory inside test folder"
-
-equiRect = cv2.imread(img_path)
-outShape = [250, 250]
-inShape = equiRect.shape[:2]
-
-print("[1] Testing equirect2Fisheye_UCM method ...")
-fisheye = mapper.equirect2Fisheye_UCM(equiRect, outShape=outShape, xi=0.5)
-img_path = "./outputs/UCM_out.jpg"
-assert os.path.isfile(img_path), "Output reference image not present"
-assert fisheye.shape[:2] == (
-    250,
-    250,
-), "Output dimensions for generated fisheye image do not match !"
-
-print("[2] Testing equirect2Fisheye_EUCM method ...")
-fisheye = mapper.equirect2Fisheye_EUCM(
-    equiRect, outShape=[250, 250], f=100, a_=0.4, b_=2, angles=[0, 0, 0]
-)
-img_path = "./outputs/EUCM_out.jpg"
-assert os.path.isfile(img_path), "Output reference image not present"
-assert fisheye.shape[:2] == (
-    250,
-    250,
-), "Output dimensions for generated fisheye image do not match !"
-
-print("[3] Testing equirect2Fisheye_FOV method ...")
-fisheye = mapper.equirect2Fisheye_FOV(
-    equiRect, outShape=[250, 250], f=40, w_=0.5, angles=[0, 0, 0]
-)
-img_path = "./outputs/FOV_out.jpg"
-assert os.path.isfile(img_path), "Output reference image not present"
-assert fisheye.shape[:2] == (
-    250,
-    250,
-), "Output dimensions for generated fisheye image do not match !"
-
-print("[4] Testing equirect2Fisheye_DS method ...")
-fisheye = mapper.equirect2Fisheye_DS(
-    equiRect, outShape=[250, 250], f=90, a_=0.4, xi_=0.8, angles=[0, 0, 0]
-)
-img_path = "./outputs/DS_out.jpg"
-assert os.path.isfile(img_path), "Output reference image not present"
-assert fisheye.shape[:2] == (
-    250,
-    250,
-), "Output dimensions for generated fisheye image do not match !"
-
-param_file_path = "./input/fisheyeParams.txt"
-assert os.path.isfile(param_file_path), "fisheyeParams.txt file not present"
-
-img_path = "./input/fisheye.jpg"
-assert os.path.isfile(img_path), "fisheye.jpg image not present"
-
-fisheye = cv2.imread(img_path)
-outShape = [200, 400]
-inShape = fisheye.shape[:2]
-
-print("[5] Testing fisheye2equirect method for mode=1...")
-mapper = fisheyeImgConv(param_file_path)
-equiRect = mapper.fisheye2equirect(fisheye, outShape)
-img_path = "./outputs/f2e_out.jpg"
-assert os.path.isfile(img_path), "Output reference image not present"
-assert equiRect.shape[:2] == (
-    200,
-    400,
-), "Output dimensions for generated fisheye image do not match !"
-
-# Mode 2
-print("[6] Testing fisheye2equirect method for mode=2...")
-equiRect = mapper.fisheye2equirect(fisheye, outShape)
-cubemap = mapper.equirect2cubemap(equiRect, modif=True)
-equiRect = mapper.cubemap2equirect(cubemap, outShape)
[…]
