### 2020-10-02 be71029
Delete tests_visual.py
