### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#17 @engnadeau
Makefile should be at root
Per the instructions in [`Running Tests`](https://github.com/kaustubh-sadekar/OmniCV-Lib#running-tests), the `Makefile` is located in `OmniCV-Lib/omnicv/`. As Makefiles are typically an entrypoint, it should be located in the repo root (i.e., `OmniCV-Lib/`)

cc https://github.com/openjournals/joss-reviews/issues/2533
