### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#18 @engnadeau
Testing and development should use a Python virtual environment
As seen in the Makefile, `PYTHON = python3` attempts to specify the Python version/executable for development and testing. This is non-standard in Python development. A virtual environment (e.g., virtualenv, pipenv, poetry) should be used instead.

cc https://github.com/openjournals/joss-reviews/issues/2533
