### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#18 @engnadeau
This also relates to the other executables that are called (e.g., flake8, pytest) as we cannot assume or control system-level installs
