### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#19 @engnadeau
Building and testing are two separate steps
Currently, the Makefile build commands (e.g., `build-py`) calls the testing (`$(MAKE) test-py`) at the end of the build script. Testing and building should be two separate steps.

Moreover, if you want to make a command that both builds and test (for convenience, but this shouldn't be a primary entrypoint), then Makefiles allow for dependent actions, e.g.:

```
build-py:
	............

test-py:
	............

build-test: build-py test-py
	............
```

In the above, `build-test` will run `build-py` and `test-py` before its own commands

cc https://github.com/openjournals/joss-reviews/issues/2533
