### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#20 @engnadeau
flake8 configurations should be in config file
Currently, the flake8 settings are in the Makefile as a one-liner. The configurations should be found in a config file (https://flake8.pycqa.org/en/latest/user/configuration.html#configuration-locations) to ensure consistent usage and better change tracking

cc https://github.com/openjournals/joss-reviews/issues/2533
