+++ omnicv/Makefile
@@ -1,8 +1,7 @@
 PYTHON = python3
-
-.PHONY = help setup test run clean
 .DEFAULT_GOAL = help
 
+.PHONY: help
 help:
 	@echo "---------------HELP-----------------"
 	@echo "To test for linting errors type make lint"
@@ -13,12 +12,15 @@ help:
 	@echo "To test only C++ extension of the project type make test-cpp"
 	@echo "To test Python and C++ extension of the project type make test-all"
 	@echo "------------------------------------"
+
+.PHONY: lint
 lint:
 	@echo "================ \n"
 	@echo "Lint with flake8 \n"
 	@echo "================ \n"
 	flake8 --per-file-ignores="__init__.py:F403,F401" && cd ../ && flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics && flake8 Examples/ && flake8 applications
 
+.PHONY: build-py
 build-py:
 	@echo "=============================================== \n"
 	@echo "Installing omnicv python library from source"
@@ -28,12 +30,14 @@ build-py:
 	@echo "=============================================== \n"
 	$(MAKE) test-py
 
+.PHONY: test-py
 test-py:
 	@echo "=============================================== \n"
 	@echo "Running tests for Python extension \n"
 	@echo "=============================================== \n"
 	cd ../tests/ && ${PYTHON} -W ignore -m pytest
 
+.PHONY: build-cpp
 build-cpp:
 	@echo "=============================================== \n"
 	@echo "Building omnicv c++ files"
@@ -41,16 +45,19 @@ build-cpp:
 	cd ../build/ && cmake -Wno-dev CMakeLists.txt && $(MAKE)
 	$(MAKE) test-cpp
 
+.PHONY: test-cpp
 test-cpp:
 	@echo "=============================================== \n"
 	@echo "Running tests for C++ extension \n"
 	@echo "=============================================== \n"
 	cd ../tests/ && ./test
 
+.PHONY: build-all
 build-all:
 	$(MAKE) build-py
 	$(MAKE) build-cpp
 
+.PHONY: test-all
 test-all:
 	$(MAKE) test-py
 	$(MAKE) test-cpp
