### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#23 @engnadeau
`sudo` should never be used
The Makefile currently calls `sudo ${PYTHON}`. Long story short, `sudo` should never be used by an external package. Moreover, this locks the repo to POSIX-based systems (i.e., Windows wouldn't work).

cc https://github.com/openjournals/joss-reviews/issues/2533
