### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#24 @engnadeau
`build/CMakeLists.txt` should be in repo root
Much like Makefiles, `CMakeLists.txt` is a type of entrypoint and should be located in the root of the repo. Typically, the workflow for CMake is e.g.:

```
cd /home/nicholas/git/OmniCV-Lib/
mkdir build
cd build
cmake ..
make
```

cc https://github.com/openjournals/joss-reviews/issues/2533
