### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#26 @engnadeau
C++ testing should use maintainable test cases and not lump all tests into a single `TEST`
Currently, `tests/test.cpp` lumps all tests into a single `TEST_CASE("Test")`. This isn't maintainable nor readable. The tests should be properly set up following best practices.

cc https://github.com/openjournals/joss-reviews/issues/2533
