### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#27 @engnadeau
Split C++ and Python code
The C++ and Python code is currently lumped into a single subfolder. Ideally, these should be in two separate subfolders for "separation of concerns", cleanliness, and maintainability. I'd suggestion `omnicv_py` and `omnicv_cpp`

cc https://github.com/openjournals/joss-reviews/issues/2533
