### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#28 @engnadeau
Code formatting should be enforced
As noted in https://github.com/kaustubh-sadekar/OmniCV-Lib/pull/10, the Python code didn't follow formatting best practices. `black` (or another formatting tool) should be enforced through a Makefile command (e.g., `make format`) and CI.

```
>>> black .
reformatted /home/nicholas/git/OmniCV-Lib/setup.py
reformatted /home/nicholas/git/OmniCV-Lib/applications/set_fisheye_params_gui.py
reformatted /home/nicholas/git/OmniCV-Lib/tests/test_.py
reformatted /home/nicholas/git/OmniCV-Lib/omnicv/omnicv.py
All done! ✨ 🍰 ✨
4 files reformatted, 10 files left unchanged.
```

cc https://github.com/openjournals/joss-reviews/issues/2533
