### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#29 @engnadeau
Has OmniCV been used in any research?
Has OmniCV been used in any research? If so, `paper/paper.md` should cite that research as it demonstrates the significance of the codebase.

cc https://github.com/openjournals/joss-reviews/issues/2533
