### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#30 @engnadeau
Remove all commented code
Related to https://github.com/kaustubh-sadekar/OmniCV-Lib/pull/13, commented code shouldn't exist in the repo

e.g. in `omnicv/omnicv.py`:

```
        # print(self.fmin)
        if np.real(self.fmin) <= 0:
            self.fmin = np.imag(self.fmin)

        # print(self.f)
        # print(self.fmin)

        mx = (x - self.Cx) / self.f
        my = (y - self.Cy) / self.f
```

cc https://github.com/openjournals/joss-reviews/issues/2533
