### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#31 @engnadeau
All public functions should have docstring
If this library is to be used and maintained by the research community, documentation is paramount. Code documentation should be enforced (e.g., by a tool like [`pydocstyle`](https://pypi.org/project/pydocstyle/):

```
>>> pydocstyle .
./setup.py:1 at module level:
        D100: Missing docstring in public module
./Examples/fisheye2equirect.py:1 at module level:
        D100: Missing docstring in public module
./Examples/cubemap2persp.py:1 at module level:
        D100: Missing docstring in public module
./Examples/equirect2persp.py:1 at module level:
        D100: Missing docstring in public module
./Examples/equirect2cubemap.py:1 at module level:
        D100: Missing docstring in public module
./Examples/equirect2fisheye.py:1 at module level:
        D100: Missing docstring in public module
./Examples/cubemap2equirect.py:1 at module level:
        D100: Missing docstring in public module
./omnicv/__init__.py:1 at module level:
        D104: Missing docstring in public package
./omnicv/omnicv.py:1 at module level:
        D100: Missing docstring in public module
./omnicv/omnicv.py:6 in public function `rmat`:
        D103: Missing docstring in public function
./omnicv/omnicv.py:33 in public class `fisheyeImgConv`:
        D101: Missing docstring in public class
./omnicv/omnicv.py:34 in public method `__init__`:
        D107: Missing docstring in __init__
./omnicv/omnicv.py:43 in public method `fisheye2equirect`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:119 in public method `equirect2cubemap`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:242 in public method `cubemap2equirect`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:341 in public method `eqruirect2persp`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:410 in public method `cubemap2persp`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:481 in public method `equirect2Fisheye`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:567 in public method `equirect2Fisheye_UCM`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:657 in public method `equirect2Fisheye_EUCM`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:759 in public method `equirect2Fisheye_FOV`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:823 in public method `equirect2Fisheye_DS`:
        D102: Missing docstring in public method
./omnicv/omnicv.py:914 in public method `applyMap`:
        D102: Missing docstring in public method
./applications/UCM_GUI.py:1 at module level:
        D100: Missing docstring in public module
./applications/UCM_GUI.py:20 in public function `nothing`:
        D103: Missing docstring in public function
./applications/set_fisheye_params_gui.py:1 at module level:
        D100: Missing docstring in public module
./applications/set_fisheye_params_gui.py:7 in public function `nothing`:
        D103: Missing docstring in public function
./applications/ViewerGUI.py:1 at module level:
        D100: Missing docstring in public module
./applications/ViewerGUI.py:7 in public function `nothing`:
        D103: Missing docstring in public function
./tests/srcpath.py:1 at module level:
        D100: Missing docstring in public module

```

cc https://github.com/openjournals/joss-reviews/issues/2533
