### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#32 @engnadeau
Better explain OmniCV and significance
Related to https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/29, the `paper/paper.md` needs to better explain the significance of the project.

When I review `omnicv/omnicv.py` and `omnicv/imgMap.cpp`, all I see is a collection of "X to Y" conversion functions.

- Why would a researcher use OmniCV over their own conversion functions (researchers typically implement their own simple algorithms)?
- What is the value proposition?

There also needs to be an example and explanation of HOW a researcher should include this repo into their project.

cc https://github.com/openjournals/joss-reviews/issues/2533
