### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#33 @engnadeau
Wrong email?
I'm assuming `@todo.todo` isn't the correct domain in `ros_files/package.xml`?

`<maintainer email="kaustubh@todo.todo">kaustubh</maintainer>`

cc https://github.com/openjournals/joss-reviews/issues/2533
