### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#33 @engnadeau
Same for the license:

```

  <license>TODO</license>
```
