### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#30 @engnadeau
Same thing in `ros_files/bkp_cmake.txt`
