### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#30 @engnadeau
and `ros_files/src/main_code.cpp`
