### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#34 @engnadeau
Duplicate code
`OmniCV-Lib/omnicv/utils.hpp` and `OmniCV-Lib/ros_files/src/utils.hpp`, `ros_files/src/imgMap.cpp` and `omnicv/imgMap.cpp` are basically identical. Why? The whole point of a library is to code once and then share. The C++ module should be imported into the ROS packages.

cc https://github.com/openjournals/joss-reviews/issues/2533
