### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#31 @engnadeau
This is also true for all the C++ functions, especially in `ros_files/src`
