### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#16 @engnadeau
@kaustubh-sadekar good start, but there's still some work to do. Please see https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/26 https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/23 for additional issues.
