### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#16 @engnadeau
Also, `pytest` should be called simply by running `pytest` in the root folder (see the documentation for proper usage: https://docs.pytest.org/en/stable/)
