### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#15 @engnadeau
Please also update `docs/_sources/index.rst.txt`
