### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#14 @engnadeau
New related issues: https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/28 https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/24 https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/23 https://github.com/kaustubh-sadekar/OmniCV-Lib/pull/22 https://github.com/kaustubh-sadekar/OmniCV-Lib/pull/21  https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/20 https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/19 https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/18
