### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#12 @engnadeau
Why is there both `.github/workflows/buildPython.yml` and `.github/workflows/testsPython.yml`? They're almost duplicates
