### 2020-10-12 kaustubh-sadekar/OmniCV-Lib#12 @engnadeau
Also, `.github/workflows/testsCpp.yml` should have a separate step for building and testing the C++. It's currently lumped into `- name: Install dependencies` which makes it impossible to read: https://github.com/kaustubh-sadekar/OmniCV-Lib/runs/1242220788?check_suite_focus=true
