### 2020-10-12 #2533 @nnadeau
@kaustubh-sadekar @mjsottile did another review pass. There are some fundamental issues to be addressed, primarily around the significance of the work. As outlined in https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/32 and https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/29, the code appears to mainly be conversion wrappers (e.g., `X to Y`). I'm not convinced that a researcher wouldn't just implement their own math/wrapper instead of pulling in an external dependency. Personally, from a software management point of view, I would need a very good value proposition of why this external project is necessary over just doing it myself.

Moreover, if the library is live up to the statements in `paper/paper.md` (e.g., `Quick and easy to use API [...]`, `Detailed documentation [...]`, `[...] easily compatible and manageable for future updates`), then there are fundamental code quality updates that are also needed. Please review and address the many issues I've submitted to the repo (https://github.com/kaustubh-sadekar/OmniCV-Lib/issues?page=1&q=author%3Annadeau). I'm especially concerned about:

- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/34
- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/31
- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/26
- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/25
- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/23
- https://github.com/kaustubh-sadekar/OmniCV-Lib/issues/18
