### 2020-10-12 #2533 @kaustubh-sadekar
It is completely true that the researchers who are proficient in writing quick and efficient code would be able to write not just the functionalities provided in the OmniCV package but any other functionalities required in their research. They would hardly need to depend on any external library. But, writing even a small piece of code, even for such researchers would utilize some of their productive time. Also, considering a more general perspective which includes researchers with intermediate proficiency in writing fast and efficient code, this library contains several functionalities that would save a fair amount of development time helping them to focus on other aspects of their research and hence promote the research process. The functionalities are provided in the two widely used programming languages in robotics research (Python and C++) and even examples for ROS nodes are provided to fulfill the same objective.  Constant and sincere efforts have been made to improve the library respecting all the suggestions provided in the reviews. After following all the valuable suggestions provided in the review process, this library will definitely save a fair amount of development time as its final updated version would contain detailed documentation, a Make file to simplify the entry point (considering a more general perspective about the coding proficiency of a user), automated testing along with continuous integration and development workflow and code which follows all the standard coding practices. Sincere efforts will be further put in to improve the library, respecting all the suggestions provided in the review process.
