### 2020-10-13 #2533 @mjsottile
Hello @kaustubh-sadekar,

Based on review feedback, my own examination of the submission, and a discussion with other editors, your submission unfortunately does not meet the threshold for acceptance into JOSS.  There are a number of rather significant issues related to software quality and the submission does not make a clear contribution to the research software community.  I hope that you take the feedback that the review process has provided and are able to refine your package.  You may consider resubmitting in the future if you feel you’ve made substantial improvements such that it would meet expectations for an acceptable JOSS publication.

Thank you @nnadeau for your patient and detailed reviews - you went above and beyond in your feedback and I greatly appreciate it.
